% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{generate_data_sim}
\alias{generate_data_sim}
\title{Generate Simulated Survival Data with Biomarkers}
\usage{
generate_data_sim(
  n = 2000,
  p = 300,
  gamma_val = 0.1,
  dt = "gompertz",
  lb_val = 1e-06,
  num_signal = 15,
  sigma_age = 10,
  beta_age = 0.1,
  beta_extra = 2,
  maxt = 10,
  age_min = 30,
  age_max = 80,
  beta_min_val = 0.05,
  beta_max_val = 0.2
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of biomarkers}

\item{num_signal}{Number of true signal biomarkers}

\item{sigma_age}{Standard deviation for age noise}

\item{beta_age}{Effect size for age}

\item{beta_extra}{Effect size for extra signal}

\item{maxt}{Maximum follow-up time}

\item{age_min}{Minimum age}

\item{age_max}{Maximum age}

\item{gamma}{Shape parameter for Gompertz distribution}

\item{lb}{Scale parameter for Gompertz distribution}

\item{beta_min}{Minimum effect size for biomarkers}

\item{beta_max}{Maximum effect size for biomarkers}
}
\value{
A data frame with simulated survival data
}
\description{
Generate Simulated Survival Data with Biomarkers
}
